refactor: move db and User into a models module inside db package
